# apps/web/src/api/openapi.example.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Periplo API (example contract)",
    "version": "0.0.0",
    "description": "Hand-derived from the HTTP contracts of the query, table discovery and ETL routes until the API publishes its own OpenAPI. Regenerate schema.d.ts with `npm run generate:api`."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/etl/status": {
      "get": {
        "operationId": "getEtlStatus",
        "responses": {
          "200": {
            "description": "Integration status; never touches the orchestrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtlStatus"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl": {
      "get": {
        "operationId": "listEtls",
        "responses": {
          "200": {
            "description": "Deployments visible to Periplo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EtlList"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/{name}/grid": {
      "get": {
        "operationId": "getEtlGrid",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The most recent non-SCHEDULED runs of the deployment as a process x run matrix, oldest to newest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunGrid"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Ambiguous deployment name (etl_ambiguous)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/{name}/runs": {
      "get": {
        "operationId": "listEtlRuns",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Runs, most recent first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Ambiguous deployment name (etl_ambiguous)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "runEtl",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Run created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "400": {
            "description": "The orchestrator rejected the request (etl_rejected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Operating ETLs is disabled (etl_operate_disabled)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Ambiguous deployment name (etl_ambiguous)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/{name}/schedule/resume": {
      "post": {
        "operationId": "resumeEtlSchedule",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Every schedule reactivated (and the deployment un-paused, if it was)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Etl"
                }
              }
            }
          },
          "400": {
            "description": "The deployment has no schedule (etl_rejected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Operating ETLs is disabled (etl_operate_disabled)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Ambiguous deployment name (etl_ambiguous)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/{name}/schedule/pause": {
      "post": {
        "operationId": "pauseEtlSchedule",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Every schedule deactivated (the inverse of loom's failure hook)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Etl"
                }
              }
            }
          },
          "400": {
            "description": "The deployment has no schedule (etl_rejected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Operating ETLs is disabled (etl_operate_disabled)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Ambiguous deployment name (etl_ambiguous)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/runs/{id}": {
      "get": {
        "operationId": "getEtlRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Run detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/runs/{id}/logs": {
      "get": {
        "operationId": "getEtlRunLogs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "The `next` of a previous page, opaque to the client; lines with timestamp >= after, ascending.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 200
            }
          },
          {
            "name": "task_run",
            "in": "query",
            "required": false,
            "description": "Repeatable, up to 100 UUIDs. Without it: flow-level lines only. With one or more: those task runs' own lines (several make up a process's scope: its marker plus its steps).",
            "schema": {
              "type": "array",
              "maxItems": 100,
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "The orchestrator's own log text search, sent as a quoted phrase; at most 200 characters.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "min_level",
            "in": "query",
            "required": false,
            "description": "A floor on log level.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log page; without `after`, the last `limit` lines",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogPage"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/runs/{id}/tasks": {
      "get": {
        "operationId": "getEtlRunTasks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run's attempts, grouped into processes and steps",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTasks"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/etl/runs/{id}/steps/{task_run}": {
      "get": {
        "operationId": "getEtlRunStep",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "task_run",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One step: its facts and the first page of its own logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepDetail"
                }
              }
            }
          },
          "404": {
            "description": "Integration off (etl_not_configured) or unknown deployment/run/step (not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "The orchestrator is unreachable or rejected the credential (etl_upstream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "Normalised client-side by @periplo/core/api; never inspected by screens.",
        "additionalProperties": true
      },
...
      "EtlStatus": {
        "type": "object",
        "required": [
          "configured",
          "operate_enabled"
        ],
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "operate_enabled": {
            "type": "boolean"
          }
        }
      },
...
      "EtlList": {
        "type": "object",
        "required": [
          "etls",
          "summary",
          "running",
          "running_truncated"
        ],
        "properties": {
          "etls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Etl"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "running": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunningRun"
            }
          },
          "running_truncated": {
            "type": "boolean",
            "description": "running was capped at 20 entries."
          }
        }
      },
      "RunList": {
        "type": "object",
        "required": [
          "runs"
        ],
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FlowRun"
            }
          }
        }
      },
      "RunDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FlowRun"
          },
          {
            "type": "object",
            "required": [
              "parameters",
              "deployment_id",
              "deployment_name",
              "flow_name",
              "terminal"
            ],
            "properties": {
              "parameters": {
                "type": "object",
                "additionalProperties": true
              },
              "deployment_id": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "deployment_name": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "flow_name": {
                "type": "string"
              },
              "terminal": {
                "type": "boolean",
                "description": "state in COMPLETED, FAILED, CANCELLED, CRASHED"
              }
            }
          }
        ]
      },
...
      "RunGrid": {
        "type": "object",
        "required": [
          "runs",
          "processes",
          "truncated"
        ],
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridRun"
            },
            "description": "Oldest to newest, left to right."
          },
          "processes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Order of first appearance in the most recent run that has any, then any others."
          },
          "truncated": {
            "type": "boolean",
            "description": "A marker, failed-step or mark query hit its hard page cap before exhausting."
          }
        }
      },
...
      "RunRequest": {
        "type": "object",
        "properties": {
          "parameters": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true,
            "description": "null runs with the deployment's own parameters."
          }
        }
      }
    }
  }
}